backend/requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0

# Database dependencies
SQLAlchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Data validation and serialization
pydantic==1.10.13
pydantic_core==2.33.2
email-validator==1.3.1

# Environment and configuration
python-dotenv==1.0.0

# HTTP and networking
httpx==0.25.2
requests==2.32.3
httpcore==1.0.9

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Production server
gunicorn==23.0.0

# File watching for development
watchfiles==1.1.0

# Additional utilities
typing_extensions==4.15.0
packaging==25.0
click==8.2.1
colorama==0.4.6